nerf-pytorch/requirement.txt
absl-py==1.1.0
astor==0.8.1
backcall==0.2.0
beautifulsoup4==4.11.1
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
clip @ git+https://github.com/openai/CLIP.git@b46f5ac7587d2e1862f8b7b1573179d80dcdd620
ConfigArgParse==1.5.3
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
filelock==3.7.1
fonttools==4.33.3
ftfy==6.1.1
gast @ file:///tmp/build/80754af9/gast_1637837472536/work
gdown==4.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio==1.46.3
h5py @ file:///tmp/build/80754af9/h5py_1637138881647/work
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.4
ipykernel==6.13.1
ipython==7.34.0
jedi==0.18.1
jupyter-client==7.3.3
jupyter-core==4.10.0
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.4.2
kornia==0.6.5
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
nest-asyncio==1.5.5
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pyzmq==23.1.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy @ file:///tmp/build/80754af9/scipy_1641536880743/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.3.2.post1
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.8.0+cu111
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt @ file:///tmp/build/80754af9/wrapt_1638420625940/work
zipp==3.8.0
